Move other non-api headers
