+++ rmw_fastrtps_dynamic_cpp/src/rmw_node.cpp
@@ -31,7 +31,8 @@
 #include "rmw_fastrtps_shared_cpp/rmw_context_impl.hpp"
 
 #include "rmw_fastrtps_dynamic_cpp/identifier.hpp"
-#include "rmw_fastrtps_dynamic_cpp/init_rmw_context_impl.hpp"
+
+#include "init_rmw_context_impl.hpp"
 
 extern "C"
 {
