+++ rmw_fastrtps_dynamic_cpp/src/rmw_subscription.cpp
@@ -31,8 +31,8 @@
 #include "rmw_fastrtps_shared_cpp/subscription.hpp"
 
 #include "rmw_fastrtps_dynamic_cpp/identifier.hpp"
-#include "rmw_fastrtps_dynamic_cpp/subscription.hpp"
 
+#include "subscription.hpp"
 #include "type_support_common.hpp"
 #include "type_support_registry.hpp"
